Formatting: column adjustment/consistency; offsets off register addressing Fixed error in Y mode of OR Added Z mode of msb_u signature fix in msb_u (s32 -> s16) Retested add modes, all routines.

+++ ui512bTests/ui512b.h
@@ -19,40 +19,45 @@ extern "C"
 
 	//	Procedures from ui512b.asm module:
 
-	// void shr_u ( u64* destination, u64* source, u32 bits_to_shift );
+	// void shr_u ( u64* destination, u64* source, u16 bits_to_shift );
 	// shift supplied source 512bit (8 QWORDS) right, put in destination
 	// EXTERNDEF	shr_u : PROC
-	void shr_u ( u64*, u64*, u32 );
+	void shr_u(u64*, u64*, u16);
 
 	// void shl_u ( u64* destination, u64* source, u16 bits_to_shift );
 	// shift supplied source 512bit (8 QWORDS) left, put in destination
 	// EXTERNDEF	shl_u : PROC
-	void shl_u ( u64*, u64*, u16 );
+	void shl_u(u64*, u64*, u16);
 
 	// void and_u ( u64* destination, u64* lh_op, u64* rh_op );
 	// logical 'AND' bits in lh_op, rh_op, put result in destination
 	// EXTERNDEF	and_u : PROC
-	void and_u ( u64*, u64*, u64* );
+	void and_u(u64*, u64*, u64*);
 
 	// void or_u ( u64* destination, u64* lh_op, u64* rh_op );
 	// logical 'OR' bits in lh_op, rh_op, put result in destination	
 	// EXTERNDEF	or_u : PROC
-	void or_u ( u64*, u64*, u64* );
+	void or_u(u64*, u64*, u64*);
 
 	// void not_u ( u64* destination, u64* source );
 	// logical 'NOT' bits in source, put result in destination
 	// EXTERNDEF	not_u : PROC
-	void not_u ( u64*, u64* );
+	void not_u(u64*, u64*);
 
-	s16 msb_u ( u64* );
+	s16 msb_u(u64*);
 	// find most significant bit in supplied source 512bit (8 QWORDS)
-	// returns: -1 if no most significant bit, bit number otherwise, bits numbered 0 to 511 inclusive
+	// returns: -1 if no most significant bit, bit number otherwise, bits numbered from left 511 to 0 inclusive
+	//	Note:	a returned zero means the significant bit is bit0 of the eighth word of the 512bit source parameter; (the right most bit)
+	//			a returned 511 means bit63 of the first word(the left most bit)
 	// EXTERNDEF	msb_u : PROC
 
-	s16 lsb_u ( u64* );
+	s16 lsb_u(u64*);
 	// find least significant bit in supplied source 512bit (8 QWORDS)
-	// returns: -1 if no least significant bit, bit number otherwise, bits numbered 0 to 511 inclusive
+	// returns: -1 if no least significant bit, bit number otherwise, bits numbered from left 511 to 0 inclusive
+	//	Note:	a returned zero means the significant bit is bit0 of the eighth word of the 512bit source parameter; (the right most bit)
+	//			a returned 511 means bit63 of the first word(the left most bit)
 	// EXTERNDEF	lsb_u : PROC
+
 };
 
 #endif
